refactor: rename banner to startup_print and move to utils
