+++ src/utils/mod.rs
@@ -1,4 +1,6 @@
 /// JSON Schema helpers for schemars.
 pub mod schema;
+/// Startup print printed to stdout when the server begins listening.
+pub mod startup_print;
 /// Unix timestamp utilities.
 pub mod time;
